feat: Add professional quant features, AI reviewer, and mobile dashboard
New modules:
- src/sentiment.rs: LunarCrush API v4 integration (limit=200 coins) for real-time market sentiment scoring
- src/ai_reviewer.rs: Anthropic Claude AI position reviewer with HOLD/EXIT/ADD recommendations
- src/coins.rs: Coin data dictionary (~120 coins) with full names and TradingView/CoinCap logos
- src/web_dashboard.rs: Mobile-first dashboard with equity hero, position cards, candidates table, signal weights strip
- src/metrics.rs: Sharpe/Sortino ratio, max drawdown, win rate, Kelly sizing calculations
- src/learner.rs: Adaptive signal weight learning from closed trade outcomes
- src/candlestick_patterns.rs: Doji, hammer, engulfing, morning/evening star pattern detection
- src/chart_patterns.rs: Head & shoulders, double top/bottom, triangles, flags detection
- deploy.sh: DigitalOcean droplet deployment script with env var forwarding

Key features:
- ADX-based regime detection (trending vs ranging markets)
- RSI/MACD/EMA indicator suite
- Confidence-scaled leverage (1x-3x based on signal confluence)
- R-multiple exits (1.5R take-profit, 0.75R stop-loss)
- Kelly criterion position sizing
- Inline signal weight strip in candidates section (compact layout)
- RedRobot logo embedded as base64 data URI in header
- Sentiment status badge (Live/No data indicator)
- AI action/reason annotations on position cards

+++ src/backtest.rs
@@ -291,7 +291,7 @@ impl Backtester {
         if self.closed_trades.is_empty() {
             0.0
         } else {
-            (winners as f64 / self.closed_trades.len()) * 100.0
+            (winners as f64 / self.closed_trades.len() as f64) * 100.0
         }
     }
